updated readme, and fixed mongo compatibility with respects to the BlogPost struct

+++ source/blog/mongo.d
@@ -28,7 +28,7 @@ BlogPost[] getPostsFromID(int id)
         //retrieve the items from the bson
         t.date = cast(string)doc["date"];
         t.name = cast(string)doc["name"];
-        t.id = cast(int)doc["id"];
+        t.id = cast(double)doc["id"];
         t.desc = cast(string)doc["desc"];
         t.content = cast(string)doc["content"];
         t.link = cast(string)doc["link"];
